# cmd/api/docs/swagger.yaml
basePath: /
definitions:
  dto.CreateRecipeRequest:
    properties:
      category:
        type: string
      cook_time:
        type: integer
      description:
        type: string
      ingredients:
        description: optional, bisa kosong
        items:
          $ref: '#/definitions/dto.IngredientInput'
        type: array
      photos:
        description: optional
        items:
          $ref: '#/definitions/dto.PhotoInput'
        type: array
      prep_time:
        type: integer
      servings:
        type: integer
      steps:
        description: optional
        items:
          $ref: '#/definitions/dto.StepInput'
        type: array
      title:
        type: string
    required:
    - title
    type: object
  dto.IngredientInput:
    properties:
      amount:
        type: string
      name:
        type: string
    required:
    - amount
    - name
    type: object
  dto.PhotoInput:
    properties:
      url:
        type: string
    required:
    - url
    type: object
  dto.StepInput:
    properties:
      detail:
        type: string
      number:
        type: integer
    required:
    - detail
    - number
    type: object
  models.Favorite:
    properties:
      id:
        type: string
      recipe:
        $ref: '#/definitions/models.Recipe'
      recipe_id:
        type: string
      user_id:
        type: string
    type: object
  models.Ingredient:
    properties:
      amount:
        type: string
      id:
        type: string
      name:
        type: string
      recipe_id:
        type: string
    type: object
  models.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  models.LoginResponse:
    properties:
      data:
        $ref: '#/definitions/models.UserResponse'
      message:
        type: string
      success:
        type: boolean
      token:
        type: string
    type: object
  models.Photo:
    properties:
      id:
        type: string
      recipe_id:
        type: string
      url:
        type: string
    type: object
  models.Recipe:
    properties:
      category:
        type: string
      cook_time:
        type: integer
      created_at:
        type: string
      description:
        type: string
      favorites:
        items:
          $ref: '#/definitions/models.Favorite'
        type: array
      id:
        type: string
      ingredients:
        items:
          $ref: '#/definitions/models.Ingredient'
        type: array
      photos:
        items:
          $ref: '#/definitions/models.Photo'
        type: array
      prep_time:
        type: integer
      servings:
        type: integer
      steps:
        items:
          $ref: '#/definitions/models.Step'
        type: array
      title:
        type: string
      updated_at:
        type: string
      user:
        allOf:
        - $ref: '#/definitions/models.User'
        description: Relations
      user_id:
        type: string
    type: object
  models.RegisterRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
    required:
    - email
    - name
    - password
    type: object
  models.RegisterResponse:
    properties:
      data:
        $ref: '#/definitions/models.UserResponse'
      message:
        type: string
      success:
        type: boolean
    type: object
  models.ResponseError:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
  models.Step:
    properties:
      detail:
        type: string
      id:
        type: string
      number:
        type: integer
      recipe_id:
        type: string
    type: object
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      favorites:
        items:
          $ref: '#/definitions/models.Favorite'
        type: array
      id:
        type: string
      name:
        type: string
      recipes:
        items:
          $ref: '#/definitions/models.Recipe'
        type: array
      updated_at:
        type: string
    type: object
  models.UserResponse:
    properties:
      email:
        type: string
      name:
        type: string
      user_id:
        type: string
    type: object
host: http://192.168.100.247:8080
info:
  contact: {}
  description: API documentation for recipe API
  title: Recipe API
  version: "1.0"
paths:
  /api/myrecipes:
    get:
      description: Get all recipes created by the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Recipe'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get my recipes
      tags:
      - Recipes
  /api/recipes:
    get:
      description: Retrieve all available recipes
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Recipe'
            type: array
      summary: Get all recipes for all users
      tags:
      - Recipes
    post:
      consumes:
      - application/json
      description: Create a new recipe with title, description, category, prep_time,
        cook_time and ingredients and steps for the logged-in user
      parameters:
      - description: Recipe Data
        in: body
        name: recipe
        required: true
        schema:
          $ref: '#/definitions/dto.CreateRecipeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Recipe'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new recipe by the authenticated user
      tags:
      - Recipes
  /api/recipes/{id}:
    delete:
      description: Delete a recipe by ID
      parameters:
      - description: Recipe ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete recipe
      tags:
      - Recipes
    get:
      description: Retrieve a recipe by its ID
      parameters:
      - description: Recipe ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Recipe'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get recipe by ID
      tags:
      - Recipes
    put:
      consumes:
      - application/json
      description: Update recipe details by ID
      parameters:
      - description: Recipe ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated Recipe
        in: body
        name: recipe
        required: true
        schema:
          $ref: '#/definitions/models.Recipe'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Recipe'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update recipe
      tags:
      - Recipes
  /api/recipes/{recipe_id}/favorites:
    delete:
      description: Explicitly remove recipe from user's favorites
      parameters:
      - description: Recipe ID
        in: path
        name: recipe_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Remove recipe from favorites
      tags:
      - Favorites
    post:
      description: Add to favorites if not exists, remove if already favorited
      parameters:
      - description: Recipe ID
        in: path
        name: recipe_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Add recipe favorite
      tags:
      - Favorites
  /api/recipes/favorites:
    get:
      consumes:
      - application/json
      description: Get all favorite recipes of the logged-in user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Favorite'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get all favorite recipes by user
      tags:
      - Favorites
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with email and password
      parameters:
      - description: Login request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Login user
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user
      parameters:
      - description: Register request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ResponseError'
      summary: Register user
      tags:
      - Auth
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"